+++ cli/commands/deploy_test.go
@@ -1,0 +1,58 @@
+package commands
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+
+	"github.com/Pallinder/go-randomdata"
+	"github.com/hasura/graphql-engine/cli/v2/internal/testutil"
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+	. "github.com/onsi/gomega/gexec"
+)
+
+var commonDeployCommandsTest = func(projectDirectory string) {
+	Context("should apply metadata, run migrations and reload metadata", func() {
+		session := testutil.Hasura(testutil.CmdOpts{
+			Args:             []string{"deploy"},
+			WorkingDirectory: projectDirectory,
+		})
+		Eventually(session, timeout).Should(Exit(0))
+		output := session.Err.Contents()
+		Expect(output).Should(ContainSubstring("Metadata applied"))
+		Expect(output).Should(ContainSubstring("migrations applied on database"))
+		Expect(output).Should(ContainSubstring("Metadata reloaded"))
+	})
+}
+
+var _ = Describe("hasura deploy", func() {
+	var projectDirectory string
+	var teardown func()
+	BeforeEach(func() {
+		projectDirectory = testutil.RandDirName()
+		hgeEndPort, teardownHGE := testutil.StartHasura(GinkgoT(), testutil.HasuraDockerImage)
+		hgeEndpoint := fmt.Sprintf("http://0.0.0.0:%s", hgeEndPort)
+
+		sourceName := randomdata.SillyName()
+		connectionString, teardownPG := testutil.StartPGContainer(GinkgoT())
+		testutil.AddPGSourceToHasura(GinkgoT(), hgeEndpoint, connectionString, sourceName)
+		copyTestConfigV3Project(projectDirectory)
+		editEndpointInConfig(filepath.Join(projectDirectory, defaultConfigFilename), hgeEndpoint)
+		editSourceNameInConfigV3ProjectTemplate(projectDirectory, sourceName, connectionString)
+
+		teardown = func() {
+			os.RemoveAll(projectDirectory)
+			teardownPG()
+			teardownHGE()
+		}
+	})
+
+	AfterEach(func() {
+		teardown()
+	})
+
+	It("deploy", func() {
+		commonDeployCommandsTest(projectDirectory)
+	})
+})