+++ cli/commands/root.go
@@ -64,6 +64,7 @@ func init() {
 		NewMetadataCmd(ec),
 		NewMigrateCmd(ec),
 		NewSeedCmd(ec),
+		NewDeployCmd(ec),
 		NewActionsCmd(ec),
 		NewPluginsCmd(ec),
 		NewVersionCmd(ec),